Cleanup dir structure
