+++ depthai_sdk/README.md
@@ -1,3 +1,23 @@
 # DepthAI SDK
 
-WIP
+DepthAI SDK is a Python package, containing convenience classes and functions that help in most common tasks while using DepthAI API.
+
+To know more about the DepthAI itself, visit [our documentation](https://docs.luxonis.com) or [our webpage](https://luxonis.com)
+
+## Installation
+
+To install this package, run the following command in your terminal window
+
+```
+$ pip install depthai-sdk
+```
+
+If you're using Raspberry Pi, providing a Pi Wheels extra package url can significantly speed up the instalation process by providing prebuilt binaries for OpenCV
+
+```
+$ pip install --extra-index-url https://www.piwheels.org/simple/ depthai-sdk
+```
+
+## Usage
+
+TBD
