+++ depthai_sdk/src/managers/blob_manager.py
