+++ depthai_sdk/src/managers/encoding_manager.py
